requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
boto3==1.5.31
botocore==1.8.45
celery==4.1.0
dj-database-url==0.4.2
Django==2.0.2
django-appconf==1.0.2
django-celery-email==2.0.0
django-filter==1.1.0
django-polymorphic==1.3
django-storages==1.6.5
django-webpack-loader==0.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
kombu==4.1.0
mysqlclient==1.3.12
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.6.0
python-dateutil==2.6.1
pytz==2018.3
redis==2.10.6
s3transfer==0.1.13
six==1.11.0
Unidecode==0.4.21
vine==1.1.4
virtualenv==15.1.0
whitenoise==3.3.1
